이메일 인증 v2 verify API 파라미터 추가 (#575)
* feat: 패스워드 변경 api v2 기능 개발

* fix: 클라이언트 페이지 딜레이상  600 -> 605초 로 협의

* fix: 코드 인증 시 이메일 유효성 제거

* fix: dto 변경

* fix: 메소드명 변경 수정

* fix: 파라미터 verifyType 추가

+++ gm-api-user/src/main/java/com/nomadconnection/dapp/api/v2/service/auth/AuthService.java
@@ -65,13 +65,15 @@ public class AuthService {
 	}
 
 	@Transactional(rollbackFor = Exception.class)
-	public ResponseEntity checkVerificationCode(String email, String code) {
+	public ResponseEntity checkVerificationCode(String email, String code, VerifyCode verifyType) {
 		String storedVerifyCode = (String) redisService.getByKey(RedisKey.VERIFICATION_CODE, email);
 
 		authValidator.expiredVerifyCodeThrowException(storedVerifyCode);
 		authValidator.mismatchedVerifyCodeThrowException(code, storedVerifyCode);
 
-		redisService.deleteByKey(RedisKey.VERIFICATION_CODE, email);
+		if(REGISTER.equals(verifyType)){
+			redisService.deleteByKey(RedisKey.VERIFICATION_CODE, email);
+		}
 		return ResponseEntity.ok().body(
 			BusinessResponse.builder().build()
         );
